Move test tool function to the test file
