+++ pkg/kubelet/util.go
@@ -16,18 +16,5 @@ limitations under the License.
 
 package kubelet
 
-import (
-	"os"
-)
-
-// dirExists returns true if the path exists and represents a directory.
-func dirExists(path string) bool {
-	s, err := os.Stat(path)
-	if err != nil {
-		return false
-	}
-	return s.IsDir()
-}
-
 // empty is a placeholder type used to implement a set
 type empty struct{}
